pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flight-price-alert"
version = "0.1.0"
description = "Scheduled flight price tracker with pluggable providers and email alerts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<0.29",
  "pydantic>=2.7,<3",
  "PyYAML>=6,<7",
]

[project.scripts]
flight-price-alert = "flight_price_alert.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
